bigdata1/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.1
coverage==7.6.12
cryptography==43.0.3
cycler==0.12.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.23.2
dvc==3.61.0
dvc-data==3.16.10
dvc-gdrive==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
fastapi==0.115.9
filelock==3.18.0
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.56.0
frozenlist==1.7.0
fsspec==2025.5.1
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grandalf==0.8
gto==1.7.2
h11==0.14.0
httplib2==0.22.0
hydra-core==1.3.2
idna==3.10
iniconfig==2.1.0
iterative-telemetry==0.0.10
joblib==1.4.2
kiwisolver==1.4.8
kombu==5.5.4
markdown-it-py==3.0.0
matplotlib==3.10.1
mdurl==0.1.2
mock==5.1.0
multidict==6.6.3
networkx==3.5
numpy==2.2.3
oauth2client==4.1.3
omegaconf==2.3.0
orjson==3.10.18
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==4.0.1
PyDrive2==1.21.3
pygit2==1.18.0
pygtrie==2.5.0
pyOpenSSL==24.2.1
pyparsing==3.2.1
pytest==8.4.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.2
scmrepo==3.3.11
seaborn==0.13.2
semver==3.0.4
setuptools==75.8.2
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
smmap==5.0.2
sniffio==1.3.1
sqltrie==0.11.2
starlette==0.45.3
tabulate==0.9.0
threadpoolctl==3.5.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.0
vine==5.1.0
voluptuous==0.15.2
wheel==0.45.1
yarl==1.20.1
zc.lockfile==3.0.post1
